+++ src/main/java/com/demo/response_entity/JournalResponse.java
@@ -5,12 +5,10 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -31,7 +29,7 @@ public class JournalResponse {
     @XmlElement(name = "dest_acc_num")
     private String destinationAccount;
 
-    private BigDecimal amount;
+    private Double amount;
     private ClientResponse payer;
     private ClientResponse recipient;
